rename Formatter classes to *.formatter.ts
