+++ src/languages/n1ql.formatter.ts
