+++ src/languages/postgresql.formatter.ts
