+++ src/languages/redshift.formatter.ts
