+++ test/redshift.test.js
@@ -1,6 +1,6 @@
 import dedent from 'dedent-js';
 import * as sqlFormatter from '../src/sqlFormatter';
-import RedshiftFormatter from '../src/languages/RedshiftFormatter';
+import RedshiftFormatter from '../src/languages/redshift.formatter';
 import behavesLikeSqlFormatter from './behavesLikeSqlFormatter';
 
 import supportsAlterTable from './features/alterTable';
